Fix issue with report catcher typewide callback with component instance specific callbacks.
Since the report_catcher is bound to uvm_report_object which is a super
type of uvm_component, it is necessary for the user component to
specify itself as a derived tyep of uvm_report_object. We don't want
users to have to do that, so code was added to check if the object
is a report object when it updates the callback queue.