fixed Mantis item 3267
Added class uvm_resource_options as a namespace to hold options that control behavior of the resources facility. Added the option "auditting" and corresponding accessor functions to turn on/off the auditting option. When turned off no audit records are added to each resource and no get records are collected in the resource pool

Modified how accessor information is collected. Previously, the accessor object was used as a key to the set of accessor records stored with each resource. This has been modified so that the name of the accessor as retrieved by get_full_name() serves as the key to the accessor map. The main benefit to this is that by not retaining the accessor handle we allow the handle to be garbage collected. By holding on to the handle garbage collected was prevented. This essentially amounts to a memory leak. Now we store only the accessor name and allow the accesor object to be garbage collected.