Checkin 1 of 2 - make process control and objections robust
The following tests are known to fail at this time on Questa, likely all three.
Most fail due to the fact the OVM semantic is preserved for the run phase.
A handful fail because the trace counts are different.

35objections/01timeout
35objections/05phases/01timeout
35objections/90tracing/01basic       -counts different
35objections/90tracing/02commandline -counts different
35objections/90tracing/03multi_objs  -counts different
40phasing/01predef
40phasing/02simple_rt
40phasing/04objection
40phasing/05waitphase
40phasing/10domains/01twodomains
40phasing/10domains/02reusedom
40phasing/20sequences/05config
40phasing/30processes/01multicomp
40phasing/30processes/02persist_mode/01simple
40phasing/30processes/02persist_mode/02fj_persist
40phasing/40jump/00nojump
40phasing/40jump/01backward
40phasing/40jump/02forward
40phasing/40jump/03nonlocal
40phasing/50phase_controls/00basic
40phasing/50phase_controls/01simple
40phasing/50phase_controls/03jump
40phasing/70user_sched/01new_sched
40phasing/70user_sched/02update_sched
40phasing/70user_sched/03set_imp
40phasing/70user_sched/04new_with_orig
40phasing/90Mantis/3301phase_ended
40phasing/90Mantis/3316domsync
50cmdlineproc/60UVM_OBJECTION_TRACE/05on -counts different

A plusarg or define (currently a define: USE_OVM_RUN_PHASE_SEMANTIC in uvm_macros.svh) controls the setting

Please qualify this check in on IUS and VCS.

A second checkin tomorrow will fix most other phase mantis items.

Checkin 2 of 2  - revised domain/sync API changes